Rename EditRequirement to EditSingleRequirement
I'm about to introduce a RequirementEditor that supports all the requirement edits, so I renamed this to disambiguate.
